+++ CHANGES.rst
@@ -8,6 +8,11 @@ CHANGES
 
 - Support subpaths and regex when parsing paths (#68).
 
+**Api**
+
+- ``_extract_path_from_service``, now returns the path name along with the path
+  swagger object (#68).
+
 
 0.5.0 (2017-02-14)
 ------------------